refactor: :label: Changing latitude and longitude to Decimal type

+++ app/domain/models/marker_model.py
@@ -1,11 +1,14 @@
+from decimal import Decimal
+
+
 class MarkerModel:
     def __init__(
         self,
         marker_id: str,
         incident_id: str,
         direction: str,
-        latitude: float,
-        longitude: float      
+        latitude: Decimal,
+        longitude: Decimal,      
     ) -> None:
         self.marker_id = marker_id
         self.incident_id = incident_id
