+++ app/infrastructure/dto/marker_request_dto.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from uuid import UUID
 from pydantic import BaseModel
 
@@ -5,5 +6,5 @@ from pydantic import BaseModel
 class MarkerRequestDTO(BaseModel):
     incident_id: UUID
     direction: str
-    latitude: float
-    longitude: float 
+    latitude: Decimal
+    longitude: Decimal
